+++ lua/fff/picker_ui/init.lua
@@ -1,0 +1,600 @@
+local M = {}
+
+local conf = require('fff.conf')
+local file_picker = require('fff.file_picker')
+local preview = require('fff.file_picker.preview')
+local utils = require('fff.utils')
+local location_utils = require('fff.location_utils')
+local layout = require('fff.layout')
+local state_manager = require('fff.picker_ui.state_manager')
+local ui_creator = require('fff.picker_ui.ui_creator')
+local search_manager = require('fff.picker_ui.search_manager')
+local renderer = require('fff.picker_ui.renderer')
+local preview_manager = require('fff.picker_ui.preview_manager')
+local navigation = require('fff.picker_ui.navigation')
+local layout_manager = require('fff.picker_ui.layout_manager')
+
+local canonicalize_fff_path = utils.canonicalize_fff_path
+
+local preview_config = conf.get().preview
+if preview_config then preview.setup(preview_config) end
+
+local function get_prompt_position() return layout.resolve_prompt_position(M.state.config) end
+
+-- Wire state from state_manager module
+M.state = state_manager.state
+
+-- Alias pure state functions from state_manager
+M.clear_selections = state_manager.clear_selections
+M.reset_history_state = state_manager.reset_history_state
+
+-- Wire ui_creator module (UI creation, window/buffer/keymap setup)
+ui_creator.init(M)
+M.create_ui = ui_creator.create_ui
+M.setup_buffers = ui_creator.setup_buffers
+M.setup_windows = ui_creator.setup_windows
+M.setup_keymaps = ui_creator.setup_keymaps
+M.focus_list_win = ui_creator.focus_list_win
+M.focus_preview_win = ui_creator.focus_preview_win
+M.focus_input_win = ui_creator.focus_input_win
+M.open_preview = ui_creator.open_preview
+M.close_preview = ui_creator.close_preview
+
+-- Wire search_manager module (search, pagination, history)
+search_manager.init(M)
+M.update_results_sync = search_manager.update_results_sync
+M.update_results = search_manager.update_results
+M.load_page_at_index = search_manager.load_page_at_index
+M.load_next_page = search_manager.load_next_page
+M.load_previous_page = search_manager.load_previous_page
+M.on_input_change = search_manager.on_input_change
+M.cycle_grep_modes = search_manager.cycle_grep_modes
+M.recall_query_from_history = search_manager.recall_query_from_history
+M.cycle_forward_query = search_manager.cycle_forward_query
+M.get_suggestion_renderer = search_manager.get_suggestion_renderer
+
+-- Wire renderer module (list rendering, scroll, empty state)
+renderer.init(M)
+M.render_list = renderer.render_list
+
+-- Wire preview_manager module (preview rendering, debounce, clear)
+preview_manager.init(M)
+M.update_preview_debounced = preview_manager.update_preview_debounced
+M.update_preview_smart = preview_manager.update_preview_smart
+M.update_preview_title = preview_manager.update_preview_title
+M.update_preview = preview_manager.update_preview
+M.clear_preview = preview_manager.clear_preview
+
+-- Wire navigation module (cursor movement, preview scroll, pagination wraparound)
+navigation.init(M)
+M.wrap_to_first = navigation.wrap_to_first
+M.wrap_to_last = navigation.wrap_to_last
+M.move_up = navigation.move_up
+M.move_down = navigation.move_down
+M.scroll_preview_up = navigation.scroll_preview_up
+M.scroll_preview_down = navigation.scroll_preview_down
+
+-- Expose helpers used by navigation
+M.scroll_to_bottom = renderer.scroll_to_bottom
+
+-- Wire layout_manager module (relayout, close)
+layout_manager.init(M)
+M.relayout = layout_manager.relayout
+M.close = layout_manager.close
+
+function M.toggle_debug()
+  local config_changed = conf.toggle_debug()
+  if config_changed then
+    local current_query = M.state.query
+    local current_items = M.state.items
+    local current_cursor = M.state.cursor
+    -- Preserve mode-specific state across close/open cycle
+    local current_mode = M.state.mode
+    local current_renderer = M.state.renderer
+    local current_grep_mode = M.state.grep_mode
+    local current_grep_config = M.state.grep_config
+    local current_filtered_items = M.state.filtered_items
+    local current_selected_files = M.state.selected_files
+    local current_selected_items = M.state.selected_items
+
+    M.close()
+    M.open({
+      mode = current_mode,
+      renderer = current_renderer,
+      grep_config = current_grep_config,
+    })
+
+    M.state.query = current_query
+    M.state.items = current_items
+    M.state.cursor = current_cursor
+    M.state.grep_mode = current_grep_mode
+    M.state.filtered_items = current_filtered_items
+    M.state.selected_files = current_selected_files
+    M.state.selected_items = current_selected_items
+    M.render_list()
+    M.update_preview()
+    M.update_status()
+
+    vim.schedule(function()
+      if M.state.active and M.state.input_win then
+        vim.api.nvim_set_current_win(M.state.input_win)
+        vim.cmd('startinsert!')
+      end
+    end)
+  else
+    M.update_results()
+  end
+end
+
+function M.render_debounced()
+  vim.schedule(function()
+    if M.state.active then
+      M.render_list()
+      M.update_preview()
+      M.update_status()
+    end
+  end)
+end
+
+--- Update status information on the right side of input using virtual text
+function M.update_status(progress)
+  if not M.state.active or not M.state.ns_id then return end
+  local config = M.state.config
+  if config == nil then return end
+
+  if M.state.mode == 'grep' then
+    -- Determine available modes to decide if we should show the mode indicator
+    -- Use grep_config.modes if provided, otherwise fall back to global config
+    ---@diagnostic disable-next-line: undefined-field
+    local modes = (M.state.grep_config and M.state.grep_config.modes)
+      or config.grep.modes
+      or { 'plain', 'regex', 'fuzzy' }
+
+    -- When regex compilation failed and we fell back to literal search, show a warning
+    local fallback_label = nil
+    if M.state.grep_regex_fallback_error then fallback_label = 'invalid regex, using literal' end
+
+    -- If only one mode configured and no fallback error, hide the mode indicator completely
+    if #modes <= 1 and not fallback_label then
+      -- Clear any existing status and don't show anything
+      vim.api.nvim_buf_clear_namespace(M.state.input_buf, M.state.ns_id, 0, -1)
+      M.state.last_status_info = nil
+      return
+    end
+
+    local keybind = config.keymaps.cycle_grep_modes
+    -- Normalize: if it's a table of keys, use the first one for display
+    if type(keybind) == 'table' then keybind = keybind[1] or '<S-Tab>' end
+
+    local mode_labels = {
+      plain = 'plain',
+      regex = 'regex',
+      fuzzy = 'fuzzy',
+    }
+    local mode_label = mode_labels[M.state.grep_mode] or 'plain'
+    local hl
+    if M.state.grep_mode == 'plain' then
+      hl = config.hl.grep_plain_active or 'Comment'
+    elseif M.state.grep_mode == 'regex' then
+      hl = config.hl.grep_regex_active or 'DiagnosticInfo'
+    else -- fuzzy
+      hl = config.hl.grep_fuzzy_active or 'DiagnosticHint'
+    end
+
+    local cache_key = keybind .. M.state.grep_mode .. (fallback_label or '')
+    if cache_key == M.state.last_status_info then return end
+    M.state.last_status_info = cache_key
+
+    vim.api.nvim_buf_clear_namespace(M.state.input_buf, M.state.ns_id, 0, -1)
+
+    local win_width = vim.api.nvim_win_get_width(M.state.input_win)
+    local available_width = win_width - 2
+
+    if fallback_label then
+      local total_len = #fallback_label
+      local col_position = available_width - total_len
+      vim.api.nvim_buf_set_extmark(M.state.input_buf, M.state.ns_id, 0, 0, {
+        virt_text = { { fallback_label, 'DiagnosticWarn' } },
+        virt_text_win_col = col_position,
+      })
+    else
+      local total_len = #keybind + 1 + #mode_label
+      local col_position = available_width - total_len
+      vim.api.nvim_buf_set_extmark(M.state.input_buf, M.state.ns_id, 0, 0, {
+        virt_text = {
+          { keybind .. ' ', hl },
+          { mode_label, hl },
+        },
+        virt_text_win_col = col_position,
+      })
+    end
+    return
+  end
+
+  -- File picker mode: show match counts
+  local status_info
+  if progress and progress.is_scanning then
+    status_info = string.format('Indexing files %d', progress.scanned_files_count)
+  else
+    local search_metadata = file_picker.get_search_metadata()
+    if #M.state.query < 2 then
+      status_info = string.format('%d', search_metadata.total_files)
+    else
+      status_info = string.format('%d/%d', search_metadata.total_matched, search_metadata.total_files)
+    end
+  end
+
+  if status_info == M.state.last_status_info then return end
+  M.state.last_status_info = status_info
+
+  vim.api.nvim_buf_clear_namespace(M.state.input_buf, M.state.ns_id, 0, -1)
+
+  local win_width = vim.api.nvim_win_get_width(M.state.input_win)
+  local available_width = win_width - 2
+  local col_position = available_width - #status_info
+
+  vim.api.nvim_buf_set_extmark(M.state.input_buf, M.state.ns_id, 0, 0, {
+    virt_text = { { status_info, 'LineNr' } },
+    virt_text_win_col = col_position,
+  })
+end
+
+--- Check whether the given window has 'winfixbuf' enabled.
+--- pcall-guarded so this stays safe on Neovim versions that predate the option.
+local window_has_winfixbuf = utils.window_has_winfixbuf
+
+--- Find the first visible window with a normal file buffer, skipping the
+--- picker's own floats.
+--- @return number|nil Window ID of the first suitable window, or nil if none found
+local function find_suitable_window()
+  local exclude = {}
+  exclude[M.state.input_win or -1] = true
+  exclude[M.state.list_win or -1] = true
+  exclude[M.state.preview_win or -1] = true
+  exclude[M.state.file_info_win or -1] = true
+  return utils.find_suitable_window(exclude)
+end
+
+--- Toggle selection for the current item.
+--- In grep mode, selection is per-occurrence; in file mode, per-file.
+function M.toggle_select()
+  if not M.state.active then return end
+
+  local was_selected = state_manager.toggle_selection()
+
+  M.render_list()
+
+  -- only when selecting the element not deselecting
+  if not was_selected then
+    if get_prompt_position() == 'bottom' then
+      M.move_up()
+    else
+      M.move_down()
+    end
+  end
+end
+
+--- Send selected files/matches to quickfix list and close picker.
+--- Normal file mode: entries at line 1, col 1.
+--- Grep mode with selections: selected occurrences with exact line/col.
+--- Grep mode without selections: re-runs search with large limit to collect all matches.
+function M.send_to_quickfix()
+  if not M.state.active then return end
+
+  local qf_list, is_grep = state_manager.send_to_quickfix()
+  if not qf_list then return end
+
+  M.close()
+
+  vim.fn.setqflist(qf_list)
+  vim.cmd('copen')
+
+  local count = #qf_list
+  local unit = is_grep and (count == 1 and 'match' or 'matches') or (count == 1 and 'file' or 'files')
+  vim.notify(string.format('Added %d %s to quickfix list', count, unit), vim.log.levels.INFO)
+end
+
+function M.select(action)
+  if not M.state.active then return end
+
+  local items = M.state.filtered_items
+  if #items == 0 or M.state.cursor > #items then return end
+
+  ---@diagnostic disable-next-line: need-check-nil
+  local item = items[M.state.cursor]
+  if not item then return end
+
+  action = action or 'edit'
+
+  -- Anchor against the indexer's base_path (may differ from cwd), then rephrase
+  -- as cwd-relative for a nicer buffer name when possible. When outside cwd,
+  -- fnamemodify(':.') leaves the absolute path intact.
+  local abs_path = canonicalize_fff_path(item.relative_path)
+  if not abs_path then return end
+  local relative_path = vim.fn.fnamemodify(abs_path, ':.')
+  local location = M.state.location -- Capture location before closing
+  local query = M.state.query -- Capture query before closing for tracking
+  local mode = M.state.mode -- Capture mode before closing for tracking
+  local suggestion_source = M.state.suggestion_source -- Capture suggestion context
+
+  -- In grep mode (or when selecting a grep suggestion), derive location from the match item
+  local is_grep_item = mode == 'grep' or suggestion_source == 'grep'
+  if is_grep_item and item.line_number and item.line_number > 0 then
+    location = { line = item.line_number }
+    if item.col and item.col > 0 then
+      location.col = item.col + 1 -- Convert 0-based byte col to 1-based
+    end
+  end
+
+  -- Fallback: if location is nil but query has a :line suffix, parse it directly
+  if not location and query and query ~= '' then
+    local line_str = query:match(':(%d+)$')
+    if line_str then
+      local line_num = tonumber(line_str)
+      if line_num and line_num > 0 then
+        local col_and_line = query:match(':(%d+):(%d+)$')
+        if col_and_line then
+          local l, c = query:match(':(%d+):(%d+)$')
+          location = { line = tonumber(l), col = tonumber(c) }
+        else
+          location = { line = line_num }
+        end
+      end
+    end
+  end
+
+  vim.cmd('stopinsert')
+  M.close()
+
+  -- Defer file open past picker float teardown. Without this, foldexpr is not
+  -- recomputed on the new window (folds appear missing) on some platforms.
+  vim.schedule(function()
+    if action == 'edit' then
+      local current_win = vim.api.nvim_get_current_win()
+      local current_buf = vim.api.nvim_get_current_buf()
+      local current_buftype = vim.api.nvim_get_option_value('buftype', { buf = current_buf })
+      local current_buf_modifiable = vim.api.nvim_get_option_value('modifiable', { buf = current_buf })
+      local current_winfixbuf = window_has_winfixbuf(current_win)
+
+      -- If the current window can't host a new buffer (special buftype, non-modifiable,
+      -- or 'winfixbuf' locking it), retarget a suitable window or fall back to a split.
+      -- Without this, :edit raises E1513 ("Cannot switch buffer. 'winfixbuf' is enabled")
+      -- whenever the picker is invoked from a window pinned via :h winfixbuf.
+      local opened_via_split = false
+      if current_buftype ~= '' or not current_buf_modifiable or current_winfixbuf then
+        local suitable_win = find_suitable_window()
+        if suitable_win then
+          vim.api.nvim_set_current_win(suitable_win)
+        elseif current_winfixbuf then
+          vim.cmd('split ' .. vim.fn.fnameescape(relative_path))
+          opened_via_split = true
+        end
+      end
+
+      if not opened_via_split then vim.cmd('edit ' .. vim.fn.fnameescape(relative_path)) end
+    elseif action == 'split' then
+      vim.cmd('split ' .. vim.fn.fnameescape(relative_path))
+    elseif action == 'vsplit' then
+      vim.cmd('vsplit ' .. vim.fn.fnameescape(relative_path))
+    elseif action == 'tab' then
+      vim.cmd('tabedit ' .. vim.fn.fnameescape(relative_path))
+    end
+
+    if location then location_utils.jump_to_location(location) end
+
+    if query and query ~= '' then
+      local config = conf.get()
+      if config.history and config.history.enabled then
+        local fff = require('fff.core').ensure_initialized()
+        -- Track in background thread (non-blocking, handled by Rust)
+        if mode == 'grep' then
+          pcall(fff.track_grep_query, query)
+        else
+          pcall(fff.track_query_completion, query, item.relative_path)
+        end
+      end
+    end
+  end)
+end
+
+--- @return string|nil Current file cache path
+local function get_current_file_cache(base_path)
+  if not base_path then return nil end
+  local current_buf = vim.api.nvim_get_current_buf()
+  if not current_buf or not vim.api.nvim_buf_is_valid(current_buf) then return nil end
+
+  local current_file = vim.api.nvim_buf_get_name(current_buf)
+  if current_file == '' then return nil end
+
+  -- Use vim.uv.fs_stat to check if file exists and is readable
+  local stat = vim.uv.fs_stat(current_file)
+  if not stat or stat.type ~= 'file' then return nil end
+
+  local absolute_path = vim.fn.fnamemodify(current_file, ':p')
+  local resolved_abs = vim.fn.resolve(absolute_path)
+  local resolved_base = vim.fn.resolve(base_path)
+
+  -- icloud direcrtoes on macos contain a lot of special characters that break
+  -- the fnamemodify which have to escaped with %
+  local escaped_base = resolved_base:gsub('([%%^$()%.%[%]*+%-?])', '%%%1')
+  local relative_path = resolved_abs:gsub('^' .. escaped_base .. '/', '')
+  if relative_path == '' or relative_path == resolved_abs then return nil end
+  return relative_path
+end
+
+--- Helper function for common picker initialization
+--- @param opts table|nil Options passed to the picker
+--- @return table|nil, string|nil Merged configuration and base path, nil config if initialization failed
+local function initialize_picker(opts)
+  local base_path = opts and opts.cwd or vim.uv.cwd()
+
+  -- Initialize file picker if needed
+  if not file_picker.is_initialized() then
+    if not file_picker.setup() then
+      vim.notify('Failed to initialize file picker', vim.log.levels.ERROR)
+      return nil
+    end
+  end
+
+  local config = conf.get()
+  local merged_config = vim.tbl_deep_extend('force', config or {}, opts or {})
+
+  return merged_config, base_path
+end
+
+--- Helper function to open UI with optional prefetched results
+--- @param query string|nil Pre-filled query (nil for empty)
+--- @param results table|nil Pre-fetched results (nil to search normally)
+--- @param location table|nil Pre-fetched location data
+--- @param merged_config table Merged configuration
+--- @param current_file_cache string|nil Current file cache
+local function open_ui_with_state(query, results, location, merged_config, current_file_cache)
+  M.state.config = merged_config
+
+  if not M.create_ui() then
+    vim.notify('Failed to create picker UI', vim.log.levels.ERROR)
+    return false
+  end
+
+  M.state.active = true
+  M.state.current_file_cache = current_file_cache
+
+  -- Set up initial state
+  if query then
+    M.state.query = query
+    vim.api.nvim_buf_set_lines(M.state.input_buf, 0, -1, false, { M.state.config.prompt .. query })
+  else
+    M.state.query = ''
+  end
+
+  if results then
+    -- Use prefetched results
+    M.state.items = results
+    M.state.filtered_items = results
+    M.state.cursor = #results > 0 and 1 or 1
+    M.state.location = location
+
+    M.render_list()
+    M.update_preview()
+    M.update_status()
+  else
+    M.update_results()
+    M.clear_preview()
+    M.update_status()
+  end
+
+  vim.api.nvim_set_current_win(M.state.input_win)
+
+  -- Position cursor at end of query if there is one
+  if query then
+    vim.schedule(function()
+      if M.state.active and M.state.input_win and vim.api.nvim_win_is_valid(M.state.input_win) then
+        vim.api.nvim_win_set_cursor(M.state.input_win, { 1, #M.state.config.prompt + #query })
+        vim.cmd('startinsert!')
+      end
+    end)
+  else
+    vim.cmd('startinsert!')
+  end
+
+  M.monitor_scan_progress(0)
+  return true
+end
+
+--- Execute a search query with callback handling before potentially opening the UI
+--- @param query string The search query to execute
+--- @param callback function Function called with results: function(results, metadata, location, get_file_score) -> boolean
+--- @param opts? table Optional configuration to override defaults (same as M.open)
+--- @return boolean true if callback handled results, false if UI was opened
+function M.open_with_callback(query, callback, opts)
+  if M.state.active then return false end
+
+  -- open_with_callback runs the file-picker flow, never grep. Reset the
+  -- renderer/mode/grep_config defensively so we can't inherit stale state
+  -- from a previous live_grep session (close() must always do this too,
+  -- but belt-and-braces).
+  M.state.renderer = nil
+  M.state.mode = nil
+  M.state.grep_config = nil
+
+  local merged_config, base_path = initialize_picker(opts)
+  if not merged_config then return false end
+
+  local current_file_cache = get_current_file_cache(base_path)
+  local results = file_picker.search_files(query, current_file_cache, nil, nil, nil)
+
+  local metadata = file_picker.get_search_metadata()
+  local location = file_picker.get_search_location()
+
+  local callback_handled = false
+  if type(callback) == 'function' then
+    local ok, result = pcall(callback, results, metadata, location, file_picker.get_file_score)
+    if ok then
+      callback_handled = result == true
+    else
+      vim.notify('Error in search callback: ' .. tostring(result), vim.log.levels.ERROR)
+    end
+  end
+
+  if callback_handled then return true end
+  open_ui_with_state(query, results, location, merged_config, current_file_cache)
+
+  return false
+end
+
+--- Open the file picker UI
+--- @param opts? {cwd?: string, title?: string, prompt?: string, max_results?: number, max_threads?: number, layout?: {width?: number|function, height?: number|function, prompt_position?: string|function, preview_position?: string|function, preview_size?: number|function}, renderer?: table, mode?: string, grep_config?: table, query?: string} Optional configuration to override defaults
+function M.open(opts)
+  if M.state.active then return end
+
+  M.state.selected_files = {}
+  M.state.selected_items = {}
+  M.state.renderer = opts and opts.renderer or nil
+  M.state.mode = opts and opts.mode or nil
+  M.state.grep_config = opts and opts.grep_config or nil
+
+  local merged_config, base_path = initialize_picker(opts)
+  if not merged_config then return end
+
+  if base_path then require('fff.core').change_indexing_directory(base_path) end
+
+  -- Initialize grep_mode to first configured mode when opening in grep mode
+  if M.state.mode == 'grep' then
+    -- Use grep_config.modes if provided, otherwise fall back to global config
+    ---@diagnostic disable-next-line: undefined-field
+    local modes = (M.state.grep_config and M.state.grep_config.modes)
+      or merged_config.grep.modes
+      or { 'plain', 'regex', 'fuzzy' }
+    M.state.grep_mode = modes[1] or 'plain'
+  end
+
+  local current_file_cache = get_current_file_cache(base_path)
+  local query = opts and opts.query or nil ---@type string|nil
+  return open_ui_with_state(query, nil, nil, merged_config, current_file_cache)
+end
+
+function M.monitor_scan_progress(iteration)
+  if not M.state.active then return end
+
+  local progress = file_picker.get_scan_progress()
+
+  if progress.is_scanning then
+    M.update_status(progress)
+
+    -- progressive decay for larger directories
+    local timeout
+    if iteration < 10 then
+      timeout = 100
+    elseif iteration < 20 then
+      timeout = 300
+    else
+      timeout = 500
+    end
+
+    vim.defer_fn(function() M.monitor_scan_progress(iteration + 1) end, timeout)
+  else
+    M.update_results()
+  end
+end
+
+return M
